don't require survey_name or weight_name arguments
The `weight_name` variable is used only if adjusting survey weights with
`target_data`. Even in that case, though, it shouldn't be necessary. If
the `weight_name` argument to `shape` isn't specified, simply create a
variable in `item_data` named `weight` with values `1L`.

The `survey_name` variable is used only if `min_survey_filter` is
applied, so only require it in conjunction with that argument.