fix the check for missing modifier_data
When using geographic modifiers via the modifier_data argument to shape(), the
data must cover all combinations of the geo and time variables in the item
response data (individual or aggregated). Because of a bug in the validation of
the geographic data, this requirement was not strictly enforced. In some cases
a warning would appear instead of an error.