Prevent inconsistent order of XX and ZZ columns
v0.2.12 introduced an issue in which the columns of XX and ZZ could be ordered inconsistently, leading to an error when shape() validated these objects. The issue could be reproduced by passing at least two variable names to the group_names argument to shape(), in an order other than alphabetic.